backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.12
celery==5.4.0
redis==5.1.1
boto3==1.35.33
aiosmtplib==3.0.2
email-validator==2.2.0
pillow==10.4.0
python-magic==0.4.27
httpx==0.27.2
